Project7_Booking/requirement.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-rest-swagger==2.1.2
djangorestframework==3.13.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
pytz==2022.2.1
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
